backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlmodel==0.0.22
networkx==3.3
tree-sitter==0.22.0
tree-sitter-python==0.21.0
tree-sitter-javascript==0.21.0
tree-sitter-typescript==0.21.0
tree-sitter-go==0.21.0
tree-sitter-java==0.21.0
websockets==12.0
zhipuai>=2.1.0
python-multipart==0.0.9
pydantic-settings==2.3.0
pytest==8.2.0
pytest-cov==5.0.0
pytest-asyncio>=0.23.0
httpx==0.27.0
